Code Cleaning
