+++ packages/SystemUI/res/values/dimens.xml
@@ -213,52 +213,4 @@
     <dimen name="qs_tile_margin_below_icon">17dp</dimen>
     <!-- Quick Settings tile geometry: icon size -->
     <dimen name="qs_tile_icon_size">32dp</dimen>
-
-    <!-- PIE PIE PIE PIE PIE PIE PIE PIE PIE PIE PIE PIE PIE PIE -->
-    <dimen name="pie_trigger_height">12dp</dimen>
-
-    <dimen name="pie_radius_start">65dp</dimen>
-    <dimen name="pie_radius_increment">80dp</dimen>
-    <dimen name="pie_outline">1.5dp</dimen>
-    <dimen name="pie_chevron_start">240dp</dimen>
-    <dimen name="pie_chevron_start_right">240dp</dimen>
-    <dimen name="pie_chevron_increment">2dp</dimen>
-    <dimen name="pie_chevron_increment_right">2dp</dimen>
-    <dimen name="pie_battery_start">155dp</dimen>
-    <dimen name="pie_battery_increment">10dp</dimen>
-    <dimen name="pie_status_start">155dp</dimen>
-    <dimen name="pie_notifications_start">250dp</dimen>
-
-    <dimen name="pie_clock_size">58dp</dimen>
-    <dimen name="pie_clock_offset">-20dp</dimen>
-    <dimen name="pie_ampm_size">15dp</dimen>
-    <dimen name="pie_ampm_offset">-67dp</dimen>
-    <dimen name="pie_status_size">12dp</dimen>
-    <dimen name="pie_status_offset">-14dp</dimen>
-    <dimen name="pie_notification_size">13dp</dimen>
-    <dimen name="pie_notification_row_size">16dp</dimen>
-    <dimen name="pie_notification_icon_size">15dp</dimen>
-
-    <dimen name="pie_snap_radius">30dp</dimen>
-    <dimen name="pie_snap_thickness">6dp</dimen>
-
-    <dimen name="pie_touch_offset">15dp</dimen>
-    <dimen name="pie_tab_title_height">24dp</dimen>
-    <dimen name="pie_item_size">40dp</dimen>
-
-    <dimen name="pie_panel_padding">20dp</dimen>
-    <!-- PIE PIE PIE PIE PIE PIE PIE PIE PIE PIE PIE PIE PIE PIE -->
-
-    <dimen name="halo_bubble_size">75dp</dimen>
-    <dimen name="halo_number_size">24dp</dimen>
-    <dimen name="halo_number_icon_size">18dp</dimen>
-    <dimen name="halo_number_text_size">14sp</dimen>
-    <dimen name="halo_speech_text_size">16sp</dimen>
-    <dimen name="halo_status_text_size">16sp</dimen>
-    <dimen name="halo_speech_hpadding_short">4dp</dimen>
-    <dimen name="halo_speech_hpadding_wide">13dp</dimen>
-    <dimen name="halo_speech_vpadding_top">3dp</dimen>
-    <dimen name="halo_speech_vpadding_bottom">3dp</dimen>
-    <dimen name="halo_content_max_width">250dp</dimen>
-
 </resources>
