+++ packages/SystemUI/res/values/pa_dimens.xml
@@ -1,0 +1,65 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright (C) 2013 ParanoidAndroid Project
+
+     Licensed under the Apache License, Version 2.0 (the "License");
+     you may not use this file except in compliance with the License.
+     You may obtain a copy of the License at
+
+          http://www.apache.org/licenses/LICENSE-2.0
+
+     Unless required by applicable law or agreed to in writing, software
+     distributed under the License is distributed on an "AS IS" BASIS,
+     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+     See the License for the specific language governing permissions and
+     limitations under the License.
+-->
+<resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
+    <!-- PIE -->
+    <dimen name="pie_trigger_height">12dp</dimen>
+
+    <dimen name="pie_radius_start">65dp</dimen>
+    <dimen name="pie_radius_increment">80dp</dimen>
+    <dimen name="pie_outline">1.5dp</dimen>
+    <dimen name="pie_chevron_start">240dp</dimen>
+    <dimen name="pie_chevron_start_right">240dp</dimen>
+    <dimen name="pie_chevron_increment">2dp</dimen>
+    <dimen name="pie_chevron_increment_right">2dp</dimen>
+    <dimen name="pie_battery_start">155dp</dimen>
+    <dimen name="pie_battery_increment">10dp</dimen>
+    <dimen name="pie_status_start">155dp</dimen>
+    <dimen name="pie_notifications_start">250dp</dimen>
+
+    <dimen name="pie_clock_size">58dp</dimen>
+    <dimen name="pie_clock_offset">-20dp</dimen>
+    <dimen name="pie_ampm_size">15dp</dimen>
+    <dimen name="pie_ampm_offset">-67dp</dimen>
+    <dimen name="pie_status_size">12dp</dimen>
+    <dimen name="pie_status_offset">-14dp</dimen>
+    <dimen name="pie_notification_size">13dp</dimen>
+    <dimen name="pie_notification_row_size">16dp</dimen>
+    <dimen name="pie_notification_icon_size">15dp</dimen>
+
+    <dimen name="pie_snap_radius">30dp</dimen>
+    <dimen name="pie_snap_thickness">6dp</dimen>
+
+    <dimen name="pie_touch_offset">15dp</dimen>
+    <dimen name="pie_tab_title_height">24dp</dimen>
+    <dimen name="pie_item_size">40dp</dimen>
+
+    <dimen name="pie_panel_padding">20dp</dimen>
+
+    <!-- HALO -->
+
+    <dimen name="halo_bubble_size">75dp</dimen>
+    <dimen name="halo_number_size">24dp</dimen>
+    <dimen name="halo_number_icon_size">18dp</dimen>
+    <dimen name="halo_number_text_size">14sp</dimen>
+    <dimen name="halo_speech_text_size">16sp</dimen>
+    <dimen name="halo_status_text_size">16sp</dimen>
+    <dimen name="halo_speech_hpadding_short">4dp</dimen>
+    <dimen name="halo_speech_hpadding_wide">13dp</dimen>
+    <dimen name="halo_speech_vpadding_top">3dp</dimen>
+    <dimen name="halo_speech_vpadding_bottom">3dp</dimen>
+    <dimen name="halo_content_max_width">250dp</dimen>
+
+</resources>
