+++ packages/SystemUI/res/values/strings.xml
@@ -42,9 +42,6 @@
     <!-- Title shown in recents popup for inspecting an application's properties -->
     <string name="status_bar_recent_inspect_item_title">App info</string>
 
-    <!-- Title shown in recents popup for showing app in floating mode -->
-    <string name="status_bar_recent_floating_item_title">Floating mode</string>
-
     <!-- Message shown in the middle of the screen after clicking on the recent apps button
          when there are no recent apps to show. Also used for accessibility. [CHAR LIMIT=45]-->
     <string name="status_bar_no_recent_apps">No recent apps</string>
@@ -321,12 +318,6 @@
     <!-- Content description of the data connection type 3.5G for accessibility (not shown on the screen). [CHAR LIMIT=NONE] -->
     <string name="accessibility_data_connection_3.5g">3.5G</string>
 
-    <!-- Content description of the data connection type HSPA+ for accessibility (not shown on the screen). [CHAR LIMIT=NONE] -->
-    <string name="accessibility_data_connection_HP">HSPA+</string>
-    
-    <!-- Content description of the data connection type DC-HSPA+ for accessibility (not shown on the screen). [CHAR LIMIT=NONE] -->
-    <string name="accessibility_data_connection_DC">DCHSPAP</string>
-
     <!-- Content description of the data connection type 4G for accessibility (not shown on the screen). [CHAR LIMIT=NONE] -->
     <string name="accessibility_data_connection_4g">4G</string>
 
@@ -439,17 +430,11 @@
     <!-- Notification text: when GPS has found a fix [CHAR LIMIT=50] -->
     <string name="gps_notification_found_text">Location set by GPS</string>
 
-    <!-- Content description of the halo button in the notification panel for accessibility (not shown on the screen). [CHAR LIMIT=NONE] -->
-    <string name="accessibility_halo">Create notification halo.</string>
-
     <!-- Content description of the clear button in the notification panel for accessibility (not shown on the screen). [CHAR LIMIT=NONE] -->
     <string name="accessibility_clear_all">Clear all notifications.</string>
 
     <!-- Title shown in notification popup for inspecting the responsible application -->
     <string name="status_bar_notification_inspect_item_title">App info</string>
-
-    <!-- Title shown in notification popup for launching notification in floating mode -->
-    <string name="status_bar_notification_floating_item_title">Floating mode</string>
 
     <!-- [CHAR LIMIT=NONE] -->
     <string name="close_universe">Close</string>
@@ -522,90 +507,14 @@
     <string name="quick_settings_wifi_display_label">Wi-Fi Display</string>
     <!-- QuickSettings: Wifi display [CHAR LIMIT=NONE] -->
     <string name="quick_settings_wifi_display_no_connection_label">Wireless Display</string>
-    <!-- QuickSettings: Profile [CHAR LIMIT=NONE] -->
-    <string name="quick_settings_profile_label">Profile</string>
     <!-- QuickSettings: Brightness dialog title [CHAR LIMIT=NONE] -->
     <string name="quick_settings_brightness_dialog_title">Brightness</string>
     <!-- QuickSettings: Brightness dialog auto brightness button [CHAR LIMIT=NONE] -->
     <string name="quick_settings_brightness_dialog_auto_brightness_label">AUTO</string>
-    <!-- QuickSettings: Desktop Mode [CHAR LIMIT=NONE] -->
-    <string name="quick_settings_desktop_mode_label">Pie</string>
-    <!-- QuickSettings: Desktop Mode [CHAR LIMIT=NONE] -->
-    <string name="quick_settings_hybrid_label">Hybrid Props</string>
 
     <!-- Title of help text shown when the notification panel is pulled down for the very first time. [CHAR LIMIT=NONE] -->
     <string name="status_bar_help_title">Notifications appear here</string>
     <!-- Body of help text shown when the notification panel is pulled down for the very first time. [CHAR LIMIT=NONE] -->
     <string name="status_bar_help_text">Access them anytime by swiping down.\nSwipe down again for system controls.</string>
-    <string name="powerwidget_screen_timeout_toast">Screen timeout set to: %1$d %2$s</string>
 
-    <string name="quick_settings_gps">GPS</string>
-    <string name="quick_settings_gps_off">GPS off</string>
-    <string name="quick_settings_screen_sleep">Go to sleep</string>
-    <string name="quick_settings_ringer_on">Sound on</string>
-    <string name="quick_settings_ringer_off">Sound off</string>
-    <string name="quick_settings_vibration_on">Vibration on</string>
-    <string name="quick_settings_vibration_off">Vibration off</string>
-    <string name="quick_settings_label_enabled">On</string>
-    <string name="quick_settings_label_disabled">Off</string>
-    <string name="quick_settings_ringer_normal">Sound</string>
-    <string name="quick_settings_lockscreen">Lock screen</string>
-    <string name="quick_settings_network_type">Network mode</string>
-    <string name="quick_settings_report_bug">Report bug</string>
-    <string name="quick_settings_sync">Sync</string>
-    <string name="quick_settings_sync_off">Sync off</string>
-    <string name="quick_settings_torch">Torch</string>
-    <string name="quick_settings_torch_off">Torch off</string>
-    <string name="quick_settings_nfc">NFC</string>
-    <string name="quick_settings_nfc_off">NFC off</string>
-    <string name="quick_settings_screen_timeout">Timeout</string>
-    <string name="quick_settings_screen_timeout_summary">%1$d %2$s</string>
-    <string name="quick_settings_usb_tether_off_label">Disconnected</string>
-    <string name="quick_settings_usb_tether_connected_label">Tethering off</string>
-    <string name="quick_settings_usb_tether_on_label">Tethering</string>
-    <string name="quick_settings_wifiap">Wi-Fi AP</string>
-    <string name="quick_settings_wifiap_off">Wi-Fi AP off</string>
-    <string name="quick_settings_lte">LTE</string>
-    <string name="quick_settings_lte_off">LTE off</string>
-    <string name="quick_settings_volume">Volume</string>
-
-    <!-- Dialog title for navigation bar button selection -->
-    <string name="navbar_dialog_title">Choose action to assign</string>
-    <string name="navbar_home_button">Home button</string>
-    <string name="navbar_recent_button">Recent button</string>
-    <string name="navbar_search_button">Search button</string>
-    <string name="navbar_back_button">Back button</string>
-    <string name="navbar_empty_button">Empty button</string>
-    <string name="navbar_menu_conditional_button">Menu (auto hide) button</string>
-    <string name="navbar_menu_always_button">Menu (always show) button</string>
-    <string name="navbar_menu_big_button">Menu button</string>
-
-    <string name="pie_date_format">ccc, dd MMM yyyy</string>
-    <string name="pie_hour_format_12">hh:mm</string>
-    <string name="pie_hour_format_24">HH:mm</string>
-    <string name="pie_am_pm">a</string>
-
-    <string name="halo_dismiss">Dismiss notification</string>
-    <string name="halo_hide">Hide</string>
-    <string name="halo_clear_all">Dismiss all notifications</string>
-    <string name="halo_silence">Hide &amp; silence until woken</string>
-    <string name="halo_tutorial1">This is HALO</string>
-    <string name="halo_tutorial2">Move me: 2x-tap &amp; drag</string>
-    <string name="halo_tutorial3">Browse tasks: Horizontal drag</string>
-    <string name="halo_battery_plugged">CHARGING</string>
-    <string name="halo_battery_unplugged">REMAINING</string>
-    <string name="halo_battery_full">FULL</string>
-    <string name="halo_aeroplane1">AEROPLANE</string>
-    <string name="halo_aeroplane2">MODE</string>
-    <string name="halo_wifi_on">WI-FI ON</string>
-    <string name="halo_wifi_off">WI-FI OFF</string>
-    <string name="halo_signal_bars_4">FULL</string>
-    <string name="halo_signal_bars_3">THREE BARS</string>
-    <string name="halo_signal_bars_2">TWO BARS</string>
-    <string name="halo_signal_bars_1">ONE BAR</string>
-    <string name="halo_signal_bars_none">NO BARS</string>
-
-
-    <!--ScreenShot Trash-->
-    <string name="trash">Trash</string>
 </resources>
